api/app/requirements.txt
Flask
Flask-GraphQL
graphene
graphene-sqlalchemy
psycopg2-binary
firebase-admin
python-dotenv
Flask-CORS
Flask-Migrate
sqlalchemy-utils
google-auth
openai>=1.0.0
google-search-results
pytz
requests
google-cloud-translate
sendgrid
sentry-sdk[flask]
